+++ doc/platforms/gbar/gpaw-venv.sh
@@ -7,14 +7,13 @@ NAME=$1
 FOLDER=$PWD
 
 echo "
+# module purge does not work as dcc-setup/2023-aug is sticky
+module unload `module -t list | grep -v 'Currently Loaded' | grep -v dcc-setup`
 source /dtu/sw/dcc/dcc-sw.bash
-module purge
-module load dcc-setup/2020-aug
-module load python/3.8.5
-unset PYTHONPATH   #Incorrectly set in python module
-module load fftw/3.3.8 libxc/4.3.4
-module load scalapack/2.1.0
-module load openblas/0.3.10
+module load python/3.11.7
+module load fftw/3.3.10 libxc/6.1.0
+module load scalapack/2.2.0
+module load openblas/0.3.28
 module load dftd3/3.2.0
 unset CC
 " > modules.sh
@@ -36,6 +35,9 @@ mv ../modules.sh bin/activate
 cat old >> bin/activate
 rm old
 
+# Fix missing dependency in preinstalled packages
+$PIP install pandas
+
 # Install ASE from git:
 git clone https://gitlab.com/ase/ase.git
 $PIP install -e ase/
@@ -53,7 +55,7 @@ fftw = True
 libraries = ['xc', 'openblas', 'fftw3', 'scalapack']
 base = Path(env['DCC_SW_PATH']) / env['DCC_SW_CPUTYPE'] / env['DCC_SW_COMPILER']
 
-for p in ['fftw/3.3.8', 'libxc/4.3.4', 'scalapack/2.1.0', 'openblas/0.3.10']:
+for p in ['fftw/3.3.10', 'libxc/6.1.0', 'scalapack/2.2.0', 'openblas/0.3.28']:
     lib = base / p / 'lib'
     library_dirs.append(lib)
     extra_link_args.append(f'-Wl,-rpath={lib}')
@@ -63,7 +65,7 @@ pip install -e . -v > gpaw.out 2>&1
 
 # Install extra basis-functions:
 cd $VENV
-export GPAW_SETUP_PATH=~jjmo/PAW/gpaw-setups-0.9.20000/
+export GPAW_SETUP_PATH=~jasc/PAW/gpaw-setups-24.11.0/
 echo "export GPAW_SETUP_PATH=$GPAW_SETUP_PATH" >> bin/activate
 
 # Tab completion: