+++ gpaw/new/calculation.py
@@ -7,7 +7,6 @@ from typing import Any, Union
 import numpy as np
 from ase import Atoms
 from ase.units import Bohr, Ha
-
 from gpaw.core import UGArray, UGDesc
 from gpaw.core.atom_arrays import AtomDistribution
 from gpaw.densities import Densities
@@ -17,6 +16,7 @@ from gpaw.mpi import broadcast as bcast
 from gpaw.mpi import broadcast_float, world
 from gpaw.new import trace, zips
 from gpaw.new.density import Density
+from gpaw.new.energies import DFTEnergies
 from gpaw.new.ibzwfs import IBZWaveFunctions
 from gpaw.new.input_parameters import InputParameters
 from gpaw.new.logger import Logger
@@ -67,7 +67,8 @@ class DFTCalculation:
                  setups: Setups,
                  scf_loop: SCFLoop,
                  pot_calc,
-                 log: Logger):
+                 log: Logger,
+                 energies: DFTEnergies | None = None):
         self.ibzwfs = ibzwfs
         self.density = density
         self.potential = potential
@@ -79,6 +80,8 @@ class DFTCalculation:
 
         self.results: dict[str, Any] = {}
         self.relpos_ac = self.pot_calc.relpos_ac
+        self.energies = energies or DFTEnergies()
+        self.forces_have_been_printed = False
 
     @classmethod
     def from_parameters(cls,
@@ -114,7 +117,7 @@ class DFTCalculation:
         scf_loop = builder.create_scf_loop()
 
         pot_calc = builder.create_potential_calculator(log)
-        potential, _ = pot_calc.calculate_without_orbitals(
+        potential, energies, _ = pot_calc.calculate_without_orbitals(
             density, kpt_band_comm=builder.communicators['D'])
         ibzwfs = builder.create_ibz_wave_functions(
             basis_set, potential, log=log)
@@ -131,7 +134,8 @@ class DFTCalculation:
         log(pot_calc)
 
         return cls(ibzwfs, density, potential,
-                   builder.setups, scf_loop, pot_calc, log)
+                   builder.setups, scf_loop, pot_calc, log,
+                   energies=energies)
 
     def move_atoms(self, atoms) -> DFTCalculation:
         check_atoms_too_close(atoms)
@@ -151,26 +155,30 @@ class DFTCalculation:
             self.density.update(self.ibzwfs)
         self.potential.move(atomdist)
 
-        new_potential, _ = self.pot_calc.calculate(
+        self.potential, self.energies, _ = self.pot_calc.calculate(
             self.density, self.ibzwfs, self.potential.vHt_x)
-        self.potential.update_from(new_potential)
 
         mm_av = self.results['non_collinear_magmoms']
         write_atoms(atoms, mm_av, self.density.nt_sR.desc, self.log)
 
         self.results = {}
+        self.forces_have_been_printed = False
 
         return self
 
     def iconverge(self, maxiter=None, calculate_forces=None):
         self.ibzwfs.make_sure_wfs_are_read_from_gpw_file()
-        yield from self.scf_loop.iterate(self.ibzwfs,
+        for ctx in self.scf_loop.iterate(self.ibzwfs,
                                          self.density,
                                          self.potential,
+                                         self.energies,
                                          self.pot_calc,
                                          maxiter=maxiter,
                                          calculate_forces=calculate_forces,
-                                         log=self.log)
+                                         log=self.log):
+            self.energies = ctx.energies
+            self.potential = ctx.potential
+            yield ctx
 
     @trace
     def converge(self,
@@ -186,26 +194,15 @@ class DFTCalculation:
         else:  # no break
             self.log('SCF steps:', step)
 
-    def energies(self):
-        energies = combine_energies(self.potential, self.ibzwfs)
+    def energy(self):
+        self.results['free_energy'] = broadcast_float(
+            self.energies.total_free, self.comm)
+        self.results['energy'] = broadcast_float(
+            self.energies.total_extrapolated, self.comm)
 
         self.log('Energy contributions relative to reference atoms:',
                  f'(reference = {self.setups.Eref * Ha:.6f})\n')
-
-        for name, e in energies.items():
-            if not name.startswith('total') and name != 'stress':
-                self.log(f'{name + ":":10}   {e * Ha:14.6f}')
-        total_free = energies['total_free']
-        total_extrapolated = energies['total_extrapolated']
-        self.log('----------------------------')
-        self.log(f'Free energy: {total_free * Ha:14.6f}')
-        self.log(f'Extrapolated:{total_extrapolated * Ha:14.6f}\n')
-
-        total_free = broadcast_float(total_free, self.comm)
-        total_extrapolated = broadcast_float(total_extrapolated, self.comm)
-
-        self.results['free_energy'] = total_free
-        self.results['energy'] = total_extrapolated
+        self.energies.summary(self.log)
 
     def dipole(self):
         if 'dipole' in self.results:
@@ -234,13 +231,15 @@ class DFTCalculation:
             self.log()
         return mm_v, mm_av
 
-    def forces(self, silent=False):
+    def forces(self):
         """Calculate atomic forces."""
-        if 'forces' not in self.results or silent:
+        if 'forces' not in self.results:
             self._calculate_forces()
 
-        if silent:
+        if self.forces_have_been_printed:
             return
+
+        self.forces_have_been_printed = True
         self.log('\nForces: [  # eV/Ang')
         F_av = self.results['forces'] * (Ha / Bohr)
         for a, setup in enumerate(self.setups):
@@ -302,7 +301,7 @@ class DFTCalculation:
         if not np.isnan(vacuum_level):
             self.log(f'vacuum-level: {vacuum_level:.3f}  # V')
             try:
-                wf1, wf2 = self.workfunctions(vacuum_level=vacuum_level)
+                wf1, wf2 = self.workfunctions(_vacuum_level=vacuum_level)
             except NonsenseError:
                 pass
             else:
@@ -310,12 +309,10 @@ class DFTCalculation:
         self.log.fd.flush()
 
     def workfunctions(self,
-                      *,
-                      vacuum_level: float | None = None
-                      ) -> tuple[float, float]:
-        if vacuum_level is None:
-            vacuum_level = self.potential.get_vacuum_level()
-        if np.isnan(vacuum_level):
+                      *, _vacuum_level=None) -> tuple[float, float]:
+        if _vacuum_level is None:
+            _vacuum_level = self.potential.get_vacuum_level()
+        if np.isnan(_vacuum_level):
             raise NonsenseError('No vacuum')
         try:
             correction = self.pot_calc.poisson_solver.dipole_layer_correction()
@@ -323,8 +320,11 @@ class DFTCalculation:
             raise NonsenseError('No dipole layer')
         correction *= Ha
         fermi_level = self.ibzwfs.fermi_level * Ha
-        wf = vacuum_level - fermi_level
+        wf = _vacuum_level - fermi_level
         return wf - correction, wf + correction
+
+    def vacuum_level(self) -> float:
+        return self.potential.get_vacuum_level()
 
     def electrostatic_potential(self) -> ElectrostaticPotential:
         return ElectrostaticPotential.from_calculation(self)
@@ -426,7 +426,7 @@ class DFTCalculation:
 
         scf_loop = builder.create_scf_loop()
         pot_calc = builder.create_potential_calculator(log)
-        potential, _ = pot_calc.calculate(density)
+        potential, energies, _ = pot_calc.calculate(density)
 
         old_ibzwfs = ibzwfs
 
@@ -456,7 +456,8 @@ class DFTCalculation:
 
         return DFTCalculation(
             ibzwfs, density, potential,
-            builder.setups, scf_loop, pot_calc, log)
+            builder.setups, scf_loop, pot_calc, log,
+            energies=energies)
 
     def get_state(self):
         return DFTState(self.ibzwfs, self.density, self.potential)
@@ -466,23 +467,6 @@ class DFTCalculation:
         warnings.warn('Use of deprecated DFTCalculation.state attribute. '
                       'Use ibzwfs, density and potential attributes instead.')
         return self.get_state()
-
-
-def combine_energies(potential: Potential,
-                     ibzwfs: IBZWaveFunctions) -> dict[str, float]:
-    """Add up energy contributions."""
-    energies = potential.energies.copy()
-    energies.pop('stress', 0.0)
-    energies['kinetic'] += ibzwfs.energies['band']
-    energies['kinetic'] += ibzwfs.energies.get('exx_kinetic', 0.0)
-    energies['xc'] += (ibzwfs.energies.get('exx_vv', 0.0) +
-                       ibzwfs.energies.get('exx_vc', 0.0) +
-                       ibzwfs.energies.get('exx_cc', 0.0))
-    energies['entropy'] = ibzwfs.energies['entropy']
-    energies['total_free'] = sum(energies.values())
-    energies['total_extrapolated'] = (energies['total_free'] +
-                                      ibzwfs.energies['extrapolation'])
-    return energies
 
 
 def write_atoms(atoms: Atoms,