+++ gpaw/new/pw/builder.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from functools import cached_property
 
+import numpy as np
 from ase.units import Ha
 
 from gpaw.core import PWDesc, UGDesc
@@ -11,10 +12,11 @@ from gpaw.core.plane_waves import PWArray
 from gpaw.new import zips
 from gpaw.new.builder import create_uniform_grid
 from gpaw.new.external_potential import create_external_potential
+from gpaw.new.gpw import as_double_precision
+from gpaw.new.pw.bloechl_poisson import BloechlPAWPoissonSolver
 from gpaw.new.pw.hamiltonian import PWHamiltonian, SpinorPWHamiltonian
 from gpaw.new.pw.hybrids import PWHybridHamiltonian
 from gpaw.new.pw.paw_poisson import SlowPAWPoissonSolver
-from gpaw.new.pw.bloechl_poisson import BloechlPAWPoissonSolver
 from gpaw.new.pw.poisson import make_poisson_solver
 from gpaw.new.pw.pot_calc import PlaneWavePotentialCalculator
 from gpaw.new.pwfd.builder import PWFDDFTComponentsBuilder
@@ -31,10 +33,12 @@ class PWDFTComponentsBuilder(PWFDDFTComponentsBuilder):
                  *,
                  comm,
                  ecut=340,
+                 dtype=None,
                  qspiral=None,
                  dedecut=None):
         self.ecut = ecut / Ha
-        super().__init__(atoms, params, comm=comm, qspiral=qspiral)
+        super().__init__(atoms, params, dtype=dtype,
+                         comm=comm, qspiral=qspiral)
 
         self._nct_ag = None
         self._tauct_ag = None
@@ -161,7 +165,9 @@ class PWDFTComponentsBuilder(PWFDDFTComponentsBuilder):
             assert self.nbands % self.communicators['b'].size == 0
             return PWHybridHamiltonian(
                 self.grid, self.wf_desc, self.xc, self.setups,
-                self.relpos_ac, self.atomdist)
+                self.relpos_ac, self.atomdist,
+                comp_charge_in_real_space=self.params.experimental.get(
+                    'ccirs'))
         return SpinorPWHamiltonian(self.qspiral_v)
 
     def convert_wave_functions_from_uniform_grid(self,
@@ -176,7 +182,7 @@ class PWDFTComponentsBuilder(PWFDDFTComponentsBuilder):
         grid = self.grid.new(kpt=kpt_c, dtype=self.dtype)
         pw = self.wf_desc.new(kpt=kpt_c)
 
-        if self.dtype == complex:
+        if np.issubdtype(self.dtype, np.complexfloating):
             emikr_R = grid.eikr(-kpt_c)
 
         mynbands, M = C_nM.dist.shape
@@ -186,7 +192,7 @@ class PWDFTComponentsBuilder(PWFDDFTComponentsBuilder):
             basis_set.lcao_to_grid(C_nM.data, psit_nR.data, q)
 
             for psit_R, psit_G in zips(psit_nR, psit_nG, strict=False):
-                if self.dtype == complex:
+                if np.issubdtype(self.dtype, np.complexfloating):
                     psit_R.data *= emikr_R
                 psit_R.fft(out=psit_G)
             return psit_nG.to_xp(self.xp)
@@ -208,6 +214,7 @@ class PWDFTComponentsBuilder(PWFDDFTComponentsBuilder):
         if 'coefficients' not in reader.wave_functions:
             return ibzwfs
 
+        singlep = reader.get('precision', 'double') == 'single'
         c = reader.bohr**1.5
         if reader.version < 0:
             c = 1  # very old gpw file
@@ -231,7 +238,7 @@ class PWDFTComponentsBuilder(PWFDDFTComponentsBuilder):
             data.scale = c
             data.length_of_last_dimension = pw.shape[-1]
 
-            if self.communicators['w'].size == 1:
+            if self.communicators['w'].size == 1 and not singlep:
                 orig_shape = data.shape
                 data.shape = shape + pw.shape
                 wfs.psit_nX = pw.from_data(data)
@@ -249,7 +256,10 @@ class PWDFTComponentsBuilder(PWFDDFTComponentsBuilder):
                 else:
                     data = [None] * (n2 - n1)
                 for psit_G, array in zips(wfs.psit_nX, data):
-                    psit_G.scatter_from(array)
+                    if singlep:
+                        psit_G.scatter_from(as_double_precision(array))
+                    else:
+                        psit_G.scatter_from(array)
 
         return ibzwfs
 
@@ -258,7 +268,7 @@ def check_g_vector_ordering(grid: UGDesc,
                             pw: PWDesc,
                             index_G: Array1D) -> None:
     size = tuple(grid.size)
-    if pw.dtype == float:
+    if np.issubdtype(pw.dtype, np.floating):
         size = (size[0], size[1], size[2] // 2 + 1)
     index0_G = pw.indices(size)
     nG = len(index0_G)