+++ gpaw/new/symmetry.py
@@ -1,17 +1,19 @@
 from __future__ import annotations
 
+from collections import defaultdict
 from functools import cached_property
-from typing import Any, Iterable
+from typing import Any, Iterable, Sequence
 
 import numpy as np
 from ase import Atoms
+from ase.units import Bohr
 from gpaw import debug
 from gpaw.core.domain import normalize_cell
 from gpaw.new import zips
 from gpaw.rotation import rotation
 from gpaw.symmetry import Symmetry as OldSymmetry
 from gpaw.symmetry import frac
-from gpaw.typing import Array3D, ArrayLike1D, ArrayLike2D, ArrayLike3D
+from gpaw.typing import Array2D, Array3D, ArrayLike1D, ArrayLike2D, ArrayLike3D
 
 
 class SymmetryBrokenError(Exception):
@@ -20,42 +22,34 @@ class SymmetryBrokenError(Exception):
 
 def create_symmetries_object(atoms: Atoms,
                              *,
-                             setup_ids=None,
-                             magmoms=None,
-                             rotations=None,
-                             translations=None,
-                             extra_ids=None,
-                             tolerance: float | None = None,
-                             point_group=True,
-                             symmorphic=True,
+                             setup_ids: Sequence | None = None,
+                             magmoms: ArrayLike2D | None = None,
+                             rotations: ArrayLike3D | None = None,
+                             translations: ArrayLike2D | None = None,
+                             atommaps: ArrayLike2D | None = None,
+                             extra_ids: Sequence[int] | None = None,
+                             tolerance: float | None = None,  # Å
+                             point_group: bool = True,
+                             symmorphic: bool = True,
                              _backwards_compatible=False) -> Symmetries:
     """Find symmetries from atoms object.
 
-    >>> sym = create_symmetries_object(Atoms('H', cell=[1, 1, 1], pbc=True))
+    >>> atoms = Atoms('H', cell=[1, 1, 1], pbc=True)
+    >>> sym = create_symmetries_object(atoms)
     >>> len(sym)
     48
     >>> sym.rotation_scc.shape
     (48, 3, 3)
     """
+    cell_cv = atoms.cell.complete()
+
     if tolerance is None:
         tolerance = 1e-7 if _backwards_compatible else 1e-5
-    if rotations is not None:
-        sym = Symmetries(cell=atoms.cell,
-                         rotations=rotations,
-                         translations=translations,
-                         tolerance=tolerance,
-                         _backwards_compatible=_backwards_compatible)
-    elif point_group:
-        sym = Symmetries.from_cell(atoms.cell.complete(),
-                                   pbc=atoms.pbc,
-                                   tolerance=tolerance,
-                                   _backwards_compatible=_backwards_compatible)
-    else:
-        sym = Symmetries(cell=atoms.cell,
-                         rotations=[[[1, 0, 0], [0, 1, 0], [0, 0, 1]]],
-                         tolerance=tolerance,
-                         _backwards_compatible=_backwards_compatible)
+    if _backwards_compatible:
+        cell_cv *= 1 / Bohr
 
+    # Create int atom-ids from setups, magmoms and user-supplied
+    # (extra_ids) ids:
     if setup_ids is None:
         ids = atoms.numbers
     else:
@@ -64,14 +58,38 @@ def create_symmetries_object(atoms: Atoms,
         ids = integer_ids((id, m) for id, m in zips(ids, safe_id(magmoms)))
     if extra_ids is not None:
         ids = integer_ids((id, x) for id, x in zips(ids, extra_ids))
-    sym = sym.new_with_positions(
-        atoms.get_scaled_positions(),
-        ids=ids,
-        symmorphic=symmorphic)
+
+    if rotations is None:
+        # Find symmetries from cell, ids and positions:
+        if point_group:
+            sym = Symmetries.from_cell(
+                cell_cv,
+                pbc=atoms.pbc,
+                tolerance=tolerance,
+                _backwards_compatible=_backwards_compatible)
+        else:
+            # No symmetries (identity only):
+            sym = Symmetries(cell=cell_cv,
+                             tolerance=tolerance,
+                             _backwards_compatible=_backwards_compatible)
+
+        sym = sym.analyze_positions(
+            atoms.get_scaled_positions(),
+            ids=ids,
+            symmorphic=symmorphic)
+    else:
+        sym = Symmetries(cell=cell_cv,
+                         rotations=rotations,
+                         translations=translations,
+                         atommaps=atommaps,
+                         tolerance=tolerance,
+                         _backwards_compatible=_backwards_compatible)
+        if atommaps is None:
+            sym = sym.with_atom_maps(atoms.get_scaled_positions(), ids=ids)
 
     # Legacy:
     sym._old_symmetry = OldSymmetry(
-        ids, atoms.cell.complete(), atoms.pbc, tolerance,
+        ids, cell_cv, atoms.pbc, tolerance,
         point_group,
         time_reversal='?',
         symmorphic=symmorphic)
@@ -79,7 +97,7 @@ def create_symmetries_object(atoms: Atoms,
     sym._old_symmetry.ft_sc = sym.translation_sc
     sym._old_symmetry.a_sa = sym.atommap_sa
     sym._old_symmetry.has_inversion = sym.has_inversion
-    sym._old_symmetry.gcd_c = sym.gcd()
+    sym._old_symmetry.gcd_c = sym.gcd_c
 
     return sym
 
@@ -88,28 +106,35 @@ class Symmetries:
     def __init__(self,
                  *,
                  cell: ArrayLike1D | ArrayLike2D,
-                 rotations: ArrayLike3D,
+                 rotations: ArrayLike3D | None = None,
                  translations: ArrayLike2D | None = None,
                  atommaps: ArrayLike2D | None = None,
-                 tolerance=1e-7,
+                 tolerance: float | None = None,
                  _backwards_compatible=False):
         """Symmetries object.
+
+        "Rotations" here means rotations, mirror and inversion operations.
+
+        Units of "cell" and "tolerance" should match.
 
         >>> sym = Symmetries.from_cell([1, 2, 3])
         >>> sym.has_inversion
         True
         >>> len(sym)
         8
-        >>> sym2 = sym.new_with_positions([[0, 0, 0], [0, 0, 0.4]],
-        ...                               ids=[1, 2])
+        >>> sym2 = sym.analyze_positions([[0, 0, 0], [0, 0, 0.4]], ids=[1, 2])
         >>> sym2.has_inversion
         False
         >>> len(sym2)
         4
         """
         self.cell_cv = normalize_cell(cell)
+        if tolerance is None:
+            tolerance = 1e-7 if _backwards_compatible else 1e-5
         self.tolerance = tolerance
         self._backwards_compatible = _backwards_compatible
+        if rotations is None:
+            rotations = [[[1, 0, 0], [0, 1, 0], [0, 0, 1]]]
         self.rotation_scc = np.array(rotations, dtype=int)
         assert (self.rotation_scc == rotations).all()
         if translations is None:
@@ -143,11 +168,13 @@ class Symmetries:
                   cell: ArrayLike1D | ArrayLike2D,
                   *,
                   pbc: ArrayLike1D = (True, True, True),
-                  tolerance=1e-7,
+                  tolerance: float | None = None,
                   _backwards_compatible=False) -> Symmetries:
         if isinstance(pbc, int):
             pbc = (pbc,) * 3
         cell_cv = normalize_cell(cell)
+        if tolerance is None:
+            tolerance = 1e-7 if _backwards_compatible else 1e-5
         rotation_scc = find_lattice_symmetry(cell_cv, pbc, tolerance,
                                              _backwards_compatible)
         return cls(cell=cell_cv,
@@ -155,29 +182,48 @@ class Symmetries:
                    tolerance=tolerance,
                    _backwards_compatible=_backwards_compatible)
 
-    def new_with_positions(self,
-                           relative_positions: ArrayLike2D | None = None,
-                           *,
-                           ids: ArrayLike1D | None = None,
-                           symmorphic: bool = True) -> Symmetries:
+    def analyze_positions(self,
+                          relative_positions: ArrayLike2D,
+                          ids: Sequence[int],
+                          *,
+                          symmorphic: bool = True) -> Symmetries:
         return prune_symmetries(
             self, np.asarray(relative_positions), ids, symmorphic)
+
+    def with_atom_maps(self,
+                       relative_positions: Array2D,
+                       ids: Sequence[int]) -> Symmetries:
+        atommap_sa = np.empty((len(self), len(relative_positions)), int)
+        a_ij = defaultdict(list)
+        for a, id in enumerate(ids):
+            a_ij[id].append(a)
+        for U_cc, t_c, map_a in zip(self.rotation_scc,
+                                    self.translation_sc,
+                                    atommap_sa):
+            map_a[:] = self.check_one_symmetry(relative_positions,
+                                               U_cc, t_c, a_ij)
+        return Symmetries(cell=self.cell_cv,
+                          rotations=self.rotation_scc,
+                          translations=self.translation_sc,
+                          atommaps=atommap_sa,
+                          tolerance=self.tolerance,
+                          _backwards_compatible=self._backwards_compatible)
 
     @classmethod
     def from_atoms(cls,
                    atoms,
                    *,
-                   ids: ArrayLike1D | None = None,
+                   ids: Sequence[int] | None = None,
                    symmorphic: bool = True,
-                   tolerance: float = 1e-7):
+                   tolerance: float | None = None):
         sym = cls.from_cell(atoms.cell,
                             pbc=atoms.pbc,
                             tolerance=tolerance)
         if ids is None:
             ids = atoms.numbers
-        return sym.new_with_positions(atoms.positions,
-                                      ids=ids,
-                                      symmorphic=symmorphic)
+        return sym.analyze_positions(atoms.positions,
+                                     ids=ids,
+                                     symmorphic=symmorphic)
 
     def __len__(self):
         return len(self.rotation_scc)
@@ -223,25 +269,21 @@ class Symmetries:
                 F_av[a2] += np.dot(F0_av[a1], op_vv)
         return F_av / len(self)
 
-    def gcd(self):
-        length_c = (self.cell_cv**2).sum(1)**0.5
-        length_c[length_c == 0.0] = 1.0
-        gcd_c = np.ones(3, int)
-        for t_c in self.translation_sc:
-            for c, t in enumerate(t_c):
-                nom, denom = frac(t, tol=self.tolerance / length_c[c])
-                if gcd_c[c] % denom != 0:
-                    gcd_c[c] *= denom
-        return gcd_c
+    def lcm(self) -> list[int]:
+        """Find least common multiple compatible with translations."""
+        return [np.lcm.reduce([frac(t, tol=1e-4)[1] for t in t_s])
+                for t_s in self.translation_sc.T]
 
     @cached_property
     def gcd_c(self):
-        return self.gcd()
+        # Needed for old gpaw.utilities.gpts.get_number_of_grid_points()
+        # function ...
+        return np.array(self.lcm())
 
     def check_grid(self, N_c) -> bool:
         """Check that symmetries are commensurate with grid."""
         for U_cc, t_c in zip(self.rotation_scc, self.translation_sc):
-            t_c *= N_c
+            t_c = t_c * N_c
             # Make sure all grid-points map onto another grid-point:
             if (((N_c * U_cc).T % N_c).any() or
                 not np.allclose(t_c, t_c.round())):
@@ -306,31 +348,31 @@ def find_lattice_symmetry(cell_cv, pbc_c, tol, _backwards_compatible=False):
     return U_scc
 
 
-def prune_symmetries(sym, spos_ac, id_a, symmorphic=True):
+def prune_symmetries(sym: Symmetries,
+                     relpos_ac: Array2D,
+                     id_a: Sequence[int],
+                     symmorphic: bool = True) -> Symmetries:
     """Remove symmetries that are not satisfied by the atoms."""
 
-    if len(spos_ac) == 0:
+    if len(relpos_ac) == 0:
         return sym
 
     # Build lists of atom numbers for each type of atom - one
     # list for each combination of atomic number, setup type,
     # magnetic moment and basis set:
-    a_ij = {}
+    a_ij = defaultdict(list)
     for a, id in enumerate(id_a):
-        if id in a_ij:
-            a_ij[id].append(a)
-        else:
-            a_ij[id] = [a]
+        a_ij[id].append(a)
 
     a_j = a_ij[id_a[0]]  # just pick the first species
 
     def check(op_cc, ft_c):
-        return sym.check_one_symmetry(spos_ac, op_cc, ft_c, a_ij)
+        return sym.check_one_symmetry(relpos_ac, op_cc, ft_c, a_ij)
 
     # if supercell disable fractional translations:
     if not symmorphic:
         op_cc = np.identity(3, int)
-        ftrans_sc = spos_ac[a_j[1:]] - spos_ac[a_j[0]]
+        ftrans_sc = relpos_ac[a_j[1:]] - relpos_ac[a_j[0]]
         ftrans_sc -= np.rint(ftrans_sc)
         for ft_c in ftrans_sc:
             a_a = check(op_cc, ft_c)
@@ -349,8 +391,8 @@ def prune_symmetries(sym, spos_ac, id_a, symmorphic=True):
             symmetries.append((op_cc, [0, 0, 0], a_a))
         elif not symmorphic:
             # check fractional translations
-            sposrot_ac = np.dot(spos_ac, op_cc)
-            ftrans_jc = sposrot_ac[a_j] - spos_ac[a_j[0]]
+            sposrot_ac = np.dot(relpos_ac, op_cc)
+            ftrans_jc = sposrot_ac[a_j] - relpos_ac[a_j[0]]
             ftrans_jc -= np.rint(ftrans_jc)
             for ft_c in ftrans_jc:
                 a_a = check(op_cc, ft_c)
@@ -359,6 +401,7 @@ def prune_symmetries(sym, spos_ac, id_a, symmorphic=True):
 
     # Add symmetry operations with fractional translations at the end:
     symmetries.extend(ftsymmetries)
+
     sym = Symmetries(cell=sym.cell_cv,
                      rotations=[s[0] for s in symmetries],
                      translations=[s[1] for s in symmetries],
@@ -366,7 +409,7 @@ def prune_symmetries(sym, spos_ac, id_a, symmorphic=True):
                      tolerance=sym.tolerance,
                      _backwards_compatible=sym._backwards_compatible)
     if debug:
-        sym.check_positions(spos_ac)
+        sym.check_positions(relpos_ac)
     return sym
 
 
@@ -507,6 +550,11 @@ def mat(rot_cc) -> str:
 
 
 def integer_ids(ids: Iterable) -> list[int]:
+    """Convert arbitrary ids to int ids.
+
+    >>> integer_ids([(1, 'a'), (12, 'b'), (1, 'a')])
+    [0, 1, 0]
+    """
     dct: dict[Any, int] = {}
     iids = []
     for id in ids: